move most logic to main.js
